SessionVar in companyHomePage
A couple of things: First, this will only hide posts if a user is logged
in. The table will still show all posts if no one is logged on. Second,
when you create a job post make sure the CompanyID you enter in the form
is the same as the User's CompanyID listed in the database.